# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/getcaptcha": {
            "post": {
                "description": "Get captcha PNG from server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get random captcha id",
                        "name": "refId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Captcha image generated"
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Respond with Token string as Json if login was succesfull.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Login input body",
                        "name": "LoginInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.login_input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successfull",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "401": {
                        "description": "username and password or captcha is not match"
                    },
                    "408": {
                        "description": "Captcha expired"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "description": "user registration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Register input body",
                        "name": "RegisterInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Register_Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "0: This Id is already registered , 1: This Username is already registered , 2: Register was successfull"
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "408": {
                        "description": "Captcha expired"
                    }
                }
            }
        },
        "/api/resetpassword": {
            "post": {
                "description": "Changes the user's password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Reset Password Body",
                        "name": "ResetPasswordInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Reset_password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "password was successfully changed"
                    },
                    "400": {
                        "description": "Empty Body"
                    },
                    "401": {
                        "description": "Invalid password for user"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/tokenValidation": {
            "post": {
                "description": "Respond with user information such as username, SSN, Role, Email and phone.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get token string",
                        "name": "BearerToken",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token is valid"
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.Register_Input": {
            "type": "object",
            "required": [
                "captchaCode",
                "captchaId",
                "email",
                "id",
                "password",
                "phone",
                "username"
            ],
            "properties": {
                "captchaCode": {
                    "type": "string"
                },
                "captchaId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "example": "0"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "controllers.Reset_password": {
            "type": "object",
            "required": [
                "newPassword",
                "oldPassword"
            ],
            "properties": {
                "newPassword": {
                    "type": "string"
                },
                "oldPassword": {
                    "type": "string"
                }
            }
        },
        "controllers.login_input": {
            "type": "object",
            "required": [
                "captchaCode",
                "captchaId",
                "pwd",
                "username"
            ],
            "properties": {
                "captchaCode": {
                    "type": "string"
                },
                "captchaId": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}